Rename order controller module
