+++ tests/test_usecase_mapping.py
@@ -3,7 +3,7 @@ from controllers.report_controller import ReportController
 from controllers.supplier_controller import SupplierController
 from controllers.component_controller import ComponentController
 from controllers.supply_controller import SupplyController
-from controllers.order_controller import OrderController
+from controllers.orders_controller import OrderController
 
 
 def test_report_controller_methods():
